added filter function to App.js
